Renamed class.
